# namviet-backend-core/docs/swagger.yaml
definitions:
...
  employment_contracts.CreateEmploymentContractRequest:
    properties:
      base_salary:
        type: number
      commission_rate_percent:
        type: number
      contract_code:
        type: string
      insurance_deduction_amount:
        type: number
      kpi_conversion_rate:
        type: number
      standard_working_days:
        type: integer
      tax_deduction_amount:
        type: number
      user_id:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - contract_code
    - user_id
    - valid_from
    type: object
  employment_contracts.EmploymentContract:
    properties:
      base_salary:
        type: number
      commission_rate_percent:
        type: number
      contract_code:
        type: string
      created_at:
        type: string
      id:
        type: string
      insurance_deduction_amount:
        type: number
      kpi_conversion_rate:
        type: number
      standard_working_days:
        type: integer
      status:
        type: string
      tax_deduction_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
  employment_contracts.UpdateEmploymentContractRequest:
    properties:
      base_salary:
        type: number
      commission_rate_percent:
        type: number
      contract_code:
        type: string
      insurance_deduction_amount:
        type: number
      kpi_conversion_rate:
        type: number
      standard_working_days:
        type: integer
      status:
        type: string
      tax_deduction_amount:
        type: number
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
...
  fund_accounts.CreateFundAccountRequest:
    properties:
      account_id:
        type: string
      account_number:
        type: string
      balance:
        type: number
      bank_id:
        type: integer
      bank_info:
        items:
          type: string
        type: array
      currency:
        type: string
      description:
        type: string
      initial_balance:
        type: number
      location:
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
  fund_accounts.FundAccount:
    properties:
      account_id:
        type: string
      account_number:
        type: string
      balance:
        type: number
      bank_id:
        type: integer
      bank_info:
        items:
          type: string
        type: array
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      initial_balance:
        type: number
      location:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  fund_accounts.UpdateFundAccountRequest:
    properties:
      account_id:
        type: string
      account_number:
        type: string
      balance:
        type: number
      bank_id:
        type: integer
      bank_info:
        items:
          type: string
        type: array
      currency:
        type: string
      description:
        type: string
      location:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for Nam Viet ERP System
  title: Nam Viet ERP API
  version: "1.0"
paths:
...
  /employment-contracts:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all employment contracts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/employment_contracts.EmploymentContract'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Employment Contracts
      tags:
      - Advanced HR
    post:
      consumes:
      - application/json
      description: Create a new employment contract
      parameters:
      - description: Contract Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employment_contracts.CreateEmploymentContractRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/employment_contracts.EmploymentContract'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Employment Contract
      tags:
      - Advanced HR
  /employment-contracts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an employment contract
      parameters:
      - description: Contract ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Employment Contract
      tags:
      - Advanced HR
    get:
      consumes:
      - application/json
      description: Retrieve a specific employment contract
      parameters:
      - description: Contract ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/employment_contracts.EmploymentContract'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Employment Contract by ID
      tags:
      - Advanced HR
    put:
      consumes:
      - application/json
      description: Update employment contract details
      parameters:
      - description: Contract ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employment_contracts.UpdateEmploymentContractRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/employment_contracts.EmploymentContract'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Employment Contract
      tags:
      - Advanced HR
...
  /fund-accounts:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all fund accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/fund_accounts.FundAccount'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Fund Accounts
      tags:
      - Fund Accounts
    post:
      consumes:
      - application/json
      description: Create a new fund account
      parameters:
      - description: Fund Account Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/fund_accounts.CreateFundAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fund_accounts.FundAccount'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Fund Account
      tags:
      - Fund Accounts
  /fund-accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a fund account
      parameters:
      - description: Fund Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Fund Account
      tags:
      - Fund Accounts
    get:
      consumes:
      - application/json
      description: Retrieve a specific fund account
      parameters:
      - description: Fund Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fund_accounts.FundAccount'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Fund Account by ID
      tags:
      - Fund Accounts
    put:
      consumes:
      - application/json
      description: Update fund account details
      parameters:
      - description: Fund Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/fund_accounts.UpdateFundAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fund_accounts.FundAccount'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Fund Account
      tags:
      - Fund Accounts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"